refactor: Move Function: fileFromDataTransferItem
